Change install instructions for LibSodium
When I ran `sudo apt-get install libsodium-dev` it didn't find it, so I changed the instructions to build the program from scratch in step 1 of installing the dependencies.  It works great, although it would be good for someone else just to make sure it's good.  It will help the less tech-savy people install this program from source.