consolidate rate limit into IRCMessageChannel.__privmsg()
All too often, I've seen takers get kicked by the IRC server for flooding,
appearing in the logs as "QUIT :RecvQ Exceeded". Most often this happens in the
stage when they send the unsigned transaction to their selected makers, and this
commit makes it so all communication to the server is sufficiently delayed to
comply with all but the most draconian of rate limits.