Update yield-generator-oscillator.py
Python 2.7 does not necessarily support 64 bit integers for type int. It's not uncommon that Python 2.7 for Windows will only allow 32 bit signed integers for type int, and that larger integers will be supported as type long.  As a result values as low as pow(2,31) = 214783648 santoshis = 21.47483648 bitcoin will be stored as a long in that case, even if one attempts to cast them as int. Changing the sanity test to allow type long seems to be the minimum change to fix this problem without creating conflicts for other users.